1-Renamed Internal Packages. 2-Abstraction Implemented
