+++ KotBasicExercise/src/main/java/com/experiments/classesandobjects/myinheritance/overridingmethods/Geometry.kt
@@ -1,0 +1,16 @@
+package com.experiments.classesandobjects.myinheritance.overridingmethods
+
+
+/*
+ As we mentioned before, we stick to making things explicit in Kotlin. So, Kotlin requires explicit modifiers for overridable members (we call them open) and for overrides
+ */
+open class Geometry {
+    open fun findArea(){}
+    open fun findPerimeter(){}
+
+    open fun justPrint() { println("---=>Geometry JustPrint method executed") }
+
+    /*open fun addition(){}
+    open fun multiplication(){}*/
+
+}
